Extract out TimeTravelingGatekeeper
Extract time-travel functions into TimeTravelingGatekeeper.sol.
This is partly because we need to remove these eventually, and
partly to make the Gatekeeper code smaller so the upgraded
gatekeeper will be deployable.

Also, deploy the TimeTravelingGatekeeper on `development` and
`ganache` networks, while deploying the regular one on all other
networks.
